+++ src/crypto_impl/node_crypto.h
@@ -19,8 +19,8 @@
 // OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE
 // USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-#ifndef SRC_NODE_CRYPTO_H_
-#define SRC_NODE_CRYPTO_H_
+#ifndef SRC_CRYPTO_IMPL_NODE_CRYPTO_H_
+#define SRC_CRYPTO_IMPL_NODE_CRYPTO_H_
 
 #if defined(NODE_WANT_INTERNALS) && NODE_WANT_INTERNALS
 
@@ -752,4 +752,4 @@ void InitCrypto(v8::Local<v8::Object> target);
 
 #endif  // defined(NODE_WANT_INTERNALS) && NODE_WANT_INTERNALS
 
-#endif  // SRC_NODE_CRYPTO_H_
+#endif  // SRC_CRYPTO_IMPL_NODE_CRYPTO_H_
