+++ src/crypto_impl/node_crypto_clienthello-inl.h
@@ -19,8 +19,8 @@
 // OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE
 // USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-#ifndef SRC_NODE_CRYPTO_CLIENTHELLO_INL_H_
-#define SRC_NODE_CRYPTO_CLIENTHELLO_INL_H_
+#ifndef SRC_CRYPTO_IMPL_NODE_CRYPTO_CLIENTHELLO_INL_H_
+#define SRC_CRYPTO_IMPL_NODE_CRYPTO_CLIENTHELLO_INL_H_
 
 #if defined(NODE_WANT_INTERNALS) && NODE_WANT_INTERNALS
 
@@ -78,4 +78,4 @@ inline bool ClientHelloParser::IsPaused() const {
 
 #endif  // defined(NODE_WANT_INTERNALS) && NODE_WANT_INTERNALS
 
-#endif  // SRC_NODE_CRYPTO_CLIENTHELLO_INL_H_
+#endif  // SRC_CRYPTO_IMPL_NODE_CRYPTO_CLIENTHELLO_INL_H_
