+++ src/crypto_impl/node_crypto_clienthello.h
@@ -19,8 +19,8 @@
 // OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE
 // USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-#ifndef SRC_NODE_CRYPTO_CLIENTHELLO_H_
-#define SRC_NODE_CRYPTO_CLIENTHELLO_H_
+#ifndef SRC_CRYPTO_IMPL_NODE_CRYPTO_CLIENTHELLO_H_
+#define SRC_CRYPTO_IMPL_NODE_CRYPTO_CLIENTHELLO_H_
 
 #if defined(NODE_WANT_INTERNALS) && NODE_WANT_INTERNALS
 
@@ -137,4 +137,4 @@ class ClientHelloParser {
 
 #endif  // defined(NODE_WANT_INTERNALS) && NODE_WANT_INTERNALS
 
-#endif  // SRC_NODE_CRYPTO_CLIENTHELLO_H_
+#endif  // SRC_CRYPTO_IMPL_NODE_CRYPTO_CLIENTHELLO_H_
