+++ src/crypto_impl/tls_wrap.h
@@ -19,8 +19,8 @@
 // OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE
 // USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-#ifndef SRC_TLS_WRAP_H_
-#define SRC_TLS_WRAP_H_
+#ifndef SRC_CRYPTO_IMPL_TLS_WRAP_H_
+#define SRC_CRYPTO_IMPL_TLS_WRAP_H_
 
 #if defined(NODE_WANT_INTERNALS) && NODE_WANT_INTERNALS
 
@@ -192,4 +192,4 @@ class TLSWrap : public AsyncWrap,
 
 #endif  // defined(NODE_WANT_INTERNALS) && NODE_WANT_INTERNALS
 
-#endif  // SRC_TLS_WRAP_H_
+#endif  //  SRC_CRYPTO_IMPL_TLS_WRAP_H_
